+++ apps/www/src/registry/default/example/playground-demo.tsx
@@ -129,6 +129,8 @@ import {
   TodoMarker,
 } from '@/registry/default/plate-ui/indent-todo-marker-component';
 
+import { ImagePreview } from '../plate-ui/image-preview';
+
 export const usePlaygroundPlugins = ({
   components = createPlateUI(),
   id,
@@ -172,7 +174,10 @@ export const usePlaygroundPlugins = ({
           createListPlugin({
             enabled: id === 'list' || !!enabled.list,
           }),
-          createImagePlugin({ enabled: !!enabled.img }),
+          createImagePlugin({
+            enabled: !!enabled.img,
+            renderAfterEditable: ImagePreview,
+          }),
           createMediaEmbedPlugin({ enabled: !!enabled.media_embed }),
           createCaptionPlugin({ ...captionPlugin, enabled: !!enabled.caption }),
           createMentionPlugin({