+++ packages/combobox/src/hooks/useHTMLInputCursorState.ts
@@ -1,23 +1,20 @@
-import { type RefObject, useCallback, useEffect, useState } from 'react';
+import { type RefObject, useCallback, useEffect, useState, useMemo } from 'react';
 
 import type { ComboboxInputCursorState } from '../types';
 
 export const useHTMLInputCursorState = (
   ref: RefObject<HTMLInputElement>
 ): ComboboxInputCursorState => {
-  const [cursorState, setCursorState] = useState<ComboboxInputCursorState>({
-    atEnd: false,
-    atStart: false,
-  });
+  const [atStart, setAtStart] = useState(false);
+  const [atEnd, setAtEnd] = useState(false);
 
   const recomputeCursorState = useCallback(() => {
-    if (!ref.current) return;
-
-    const { selectionEnd, selectionStart, value } = ref.current;
-
-    setCursorState({
-      atEnd: selectionEnd === value.length,
-      atStart: selectionStart === 0,
+    // Wait for events to finish processing
+    setTimeout(() => {
+      if (!ref.current) return;
+      const { selectionEnd, selectionStart, value } = ref.current;
+      setAtStart(selectionStart === 0);
+      setAtEnd(selectionEnd === value.length);
     });
   }, [ref]);
 
@@ -31,11 +28,25 @@ export const useHTMLInputCursorState = (
     input.addEventListener('input', recomputeCursorState);
     input.addEventListener('selectionchange', recomputeCursorState);
 
+    /**
+     * Compat: selectionchange is not supported for <input> except in Firefox,
+     * so we add keydown, pointerdown and pointerup as fallbacks (2024-06-14).
+     */
+    input.addEventListener('keydown', recomputeCursorState);
+    input.addEventListener('pointerdown', recomputeCursorState);
+    input.addEventListener('pointerup', recomputeCursorState);
+
     return () => {
       input.removeEventListener('input', recomputeCursorState);
       input.removeEventListener('selectionchange', recomputeCursorState);
+      input.removeEventListener('keydown', recomputeCursorState);
+      input.removeEventListener('pointerdown', recomputeCursorState);
+      input.removeEventListener('pointerup', recomputeCursorState);
     };
   }, [recomputeCursorState, ref]);
 
-  return cursorState;
+  return useMemo(() => ({
+    atStart,
+    atEnd,
+  }), [atStart, atEnd]);
 };