+++ packages/selection/src/utils/index.ts
@@ -6,5 +6,6 @@ export * from './copySelectedBlocks';
 export * from './extractSelectableIds';
 export * from './getAboveDomNode';
 export * from './onCloseBlockSelection';
+export * from './openContextMenu';
 export * from './pasteSelectedBlocks';
 export * from './selectInsertedBlocks';