+++ packages/caption/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @udecode/plate-caption
 
+## 34.1.0
+
+### Minor Changes
+
+- [#3289](https://github.com/udecode/plate/pull/3289) by [@felixfeng33](https://github.com/felixfeng33) – Add `showCaption` utils
+
+## 34.0.8
+
+### Patch Changes
+
+- [#3287](https://github.com/udecode/plate/pull/3287) by [@TyMick](https://github.com/TyMick) – Update caption UI when element `caption` property changes
+
 ## 34.0.3
 
 ### Patch Changes