+++ packages/math/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @udecode/plate-math
 
+## 34.1.0
+
+### Minor Changes
+
+- [#3289](https://github.com/udecode/plate/pull/3289) by [@felixfeng33](https://github.com/felixfeng33) – Add `inline-equation` toolbar button
+
 ## 34.0.0
 
 ### Minor Changes