+++ packages/slash-command/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @udecode/plate-slash-command
 
+## 34.0.8
+
 ## 34.0.0
 
 ### Major Changes